requirements.txt
python_doctr==0.9.0
